requirements.txt
anyio==4.6.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
async-lru==2.0.4
attrs==24.2.0
babel==2.16.0
beautifulsoup4==4.12.3
black==24.8.0
bleach==6.1.0
build==1.2.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
crispy-bootstrap5==2024.2
cryptography==43.0.1
debugpy==1.8.6
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20230430
Django==5.1.1
django-allauth==0.54.0
django-appconf==1.0.6
django-browser-reload==1.15.0
django-crispy-forms==2.3
django-debug-toolbar==4.4.6
django-dynamic-preferences==1.16.0
django-extensions==3.2.3
django-filter==24.3
django-formset==1.5.3
django-import-export==4.1.1
django-jazzmin==3.0.0
django-jinja==2.11.0
django-model-utils==5.0.0
django-notifications-hq==1.8.3
django-select2==8.2.1
django-simple-history==3.7.0
djangorestframework==3.15.1
drf-spectacular==0.27.2
executing==2.1.0
factory_boy==3.3.1
Faker==29.0.0
fastjsonschema==2.20.0
forex-python==1.8
fqdn==1.5.1
graphviz==0.20.3
graypy==2.1.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.10
inflection==0.5.1
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.27.0
ipywidgets==8.1.5
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.4
json5==0.9.25
jsonfield==3.1.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
loguru==0.7.2
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
mypy-extensions==1.0.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.2.2
notebook_shim==0.2.4
oauthlib==3.2.2
ofxparse==0.21
overrides==7.7.0
packaging==24.1
pandocfilters==1.5.1
parso==0.8.4
pathspec==0.12.1
persisting-theory==1.0
pexpect==4.9.0
platformdirs==4.3.6
pluggy==1.5.0
prometheus_client==0.21.0
prompt_toolkit==3.0.48
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pygelf==0.4.2
Pygments==2.18.0
PyJWT==2.9.0
pyproject_hooks==1.1.0
pytest==8.3.3
pytest-django==4.9.0
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
python-magic==0.4.27
python3-openid==3.2.0
pytz==2024.2
pywatchman==2.0.0
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.8.1
rpds-py==0.20.0
Send2Trash==1.8.3
setuptools==75.1.0
simplejson==3.19.3
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.1
stack-data==0.6.3
swapper==1.4.0
tablib==3.5.0
terminado==0.18.1
tinycss2==1.3.0
tornado==6.4.1
traitlets==5.14.3
types-python-dateutil==2.9.0.20240906
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.2.3
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.4
widgetsnbextension==4.0.13
xstate-machine==3.2.0
